Extract run_game from models/game.rs into dedicated game_runner module
The 250-line run_game function was game orchestration logic (engine init,
snake API calls, turn loop, DB persistence, placement calculation) living
in a data model file. This moves it to its own top-level module where it
belongs, keeping models/game.rs focused on the Game struct and CRUD.
