+++ server/src/main.rs
@@ -17,6 +17,7 @@ mod engine_models;
 mod errors;
 mod flasher;
 mod game_channels;
+mod game_runner;
 mod github;
 mod jobs;
 mod models;
